third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-getStats-promise.html

<!DOCTYPE html>
<html>
<head>
<title>RTCPeerConnection.getStats</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
var pc = new RTCPeerConnection();

function stringify_equals(d1, d2) {
  return JSON.stringify(d1) == JSON.stringify(d2);
}

function assert_expected_report(report) {
  assert_true(report != null);
  var statsCount = 0;
  report.forEach(function(s) {
    assert_true(s != null);
    statsCount++;
    assert_true(stringify_equals(s, report.get(s.id)));
  });

  // Verify that the report contains the same stats that
  // |MockWebRTCPeerConnectionHandler::getStats| should produce.
  assert_equals(statsCount, 1);
  var stats = report.get('mock-stats-01');
  assert_true(stats != null);
  assert_equals(stats.type, 'mock-stats');
  assert_equals(stats.timestamp, 1234.0);
  assert_equals(stats.bool, true);
  assert_equals(stats.int32, 42);
  assert_equals(stats.uint32, 42);
  assert_equals(stats.int64, 42);
  assert_equals(stats.uint64, 42);
  assert_equals(stats.double, 42);
  assert_equals(stats.string, '42');
  assert_true(stringify_equals(stats.sequenceBool, [true]));
  assert_true(stringify_equals(stats.sequenceInt32, [42]));
  assert_true(stringify_equals(stats.sequenceUint32, [42]));
  assert_true(stringify_equals(stats.sequenceInt64, [42]));
  assert_true(stringify_equals(stats.sequenceUint64, [42]));
  assert_true(stringify_equals(stats.sequenceDouble, [42]));
  assert_true(stringify_equals(stats.sequenceString, ['42']));
}

promise_test(function() {
  let test = this;
  return navigator.mediaDevices.getUserMedia({audio:true, video:true})
    .then(function(mediaStream) {
        pc.addStream(mediaStream);
        var selector = pc.getLocalStreams()[0].getVideoTracks()[0];
        assert_not_equals(selector, null);
        return promise_rejects(test, new TypeError(), pc.getStats(selector));
      });
}, 'getStats(MediaStreamTrack selector)');

promise_test(function() {
  let test = this;
  return navigator.mediaDevices.getUserMedia({audio:true, video:true})
    .then(function(mediaStream) {
        pc.addStream(mediaStream);
        return promise_rejects(test, new TypeError(), pc.getStats(null));
      });
}, 'getStats(null)');

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({audio:true, video:true})
    .then(function(mediaStream) {
        pc.addStream(mediaStream);
        return pc.getStats()
          .then(function(report) {
              assert_expected_report(report);
            });
      });
}, 'getStats()');
</script>
</body>
</html>
